pyproject.toml
[project]
name = "streamblocks"
version = "0.1.0"
description = "Real-time extraction and processing of structured blocks from text streams"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Text Processing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]


[project.urls]
"Homepage" = "https://github.com/streamblocks/streamblocks"
"Bug Tracker" = "https://github.com/streamblocks/streamblocks/issues"
"Source" = "https://github.com/streamblocks/streamblocks"
"Documentation" = "https://streamblocks.github.io/streamblocks/latest/"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/streamblocks/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/streamblocks"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "tests/",
    "CHANGELOG.md",
    "LICENSE",
    "README.md",
    "cliff.toml",
    "lefthook.yml",
    "pyproject.toml",
]


[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.21",
    "hypothesis>=6.0",
    "ruff>=0.8",
    "mypy>=1.14",
    "pretty_errors>=1.2",
    "basedpyright>=1.31.4",
    "types-pyyaml>=6.0.12.20250915",
]
doc = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-material-extensions>=1.3",
    "mike>=2.1",
    "mkdocstrings[python]>=0.27"
]


[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
# Enable multiple rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "PTH",    # flake8-use-pathlib
    "PL",     # pylint
]

# Ignore specific rules
ignore = [
    "PLR0913",  # Too many arguments
    "ISC001",   # Conflicts with formatter
]

# Configure import sorting (replaces isort)
[tool.ruff.lint.isort]
known-first-party = ["streamblocks"]

# Format settings
[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.per-file-ignores]
"tests/**/*" = ["PLR2004"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
extra_checks = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__", "**/migrations", ".venv"]

pythonVersion = "3.12"
venvPath = ".venv"

typeCheckingMode = "strict"
reportPossiblyUnboundVariable = true
reportUnnecessaryTypeIgnore = true
reportMissingTypeStubs = true
reportUnknownParameterType = "error"
reportUnknownMemberType = "error"
reportPrivateUsage = "error"
reportConstantRedefinition = "error"
reportImplicitStringConcatenation = "error"
reportCallInDefaultInitializer = "error"
reportPropertyTypeMismatch = "error"
reportUnusedImport = "error"
reportUnusedClass = "error"
reportUnusedFunction = "error"
reportUnusedVariable = "error"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "--cov=streamblocks --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source_pkgs = ["streamblocks"]
branch = true
parallel = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.hatch.envs.default]
installer = "uv"

[tool.hatch.envs.default.scripts]
release = "python tools/release.py {args}"


